New character, new animations, bug fixes and performance improvements
* New Character: The Viking!
Finally, a new character. The Viking is a somewhat fast fighter that can deal a ton of damage if used properly. He has relatively strong attacks with small time frames. His special ability is to become unstopable. While in this unstopable state, he cannot be stunned, but can still receieve damage and be killed.

* The attack debug component kinda works now, but isn't properly placed in the world which makes it hard to tell if the angle of the attack is rendered correctly.

* Fixed a bug with side stepping clipping through the stage boundaries.

* Fixed a bug that sometimes caused a round draw.

* Made FighManager code a little cleaner.

* Fixed a bug that caused player health miscalculations when the round timer reached zero which ultimately led to the incorrect player winning the round.

* Improved performance of the landscape terrain. It was really bad before, but should be much better now.