Just too many changes on Machine Learning.
It's basically impossible to list all the changes, so I'll just say this: The AI is now done and the machine learning is in a satisfactory state now. The CPU can fight pretty well and knows how to use blocking and its special moves as well.

Oh right, I also fixed a bug that caused instant defeat when side stepping and being attacked at the same time.